+++ order-service/src/main/java/com/rushcrew/order/OrderApplication.java
@@ -1,0 +1,13 @@
+package com.rushcrew.order;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OrderApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OrderApplication.class, args);
+	}
+
+}
