+++ order-service/src/main/java/com/rushcrew/order/application/event/producer/OrderEventPublisher.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.application.event.producer;
+
+public class OrderEventPublisher {
+}
