+++ order-service/src/main/java/com/rushcrew/order/application/service/OrderService.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.application.service;
+
+public class OrderService {
+}
