+++ order-service/src/main/java/com/rushcrew/order/domain/history/OrderHistoryRepository.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.history;
+
+public interface OrderHistoryRepository {
+}
