+++ order-service/src/main/java/com/rushcrew/order/domain/order/Order.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.order;
+
+public class Order {
+}
