+++ order-service/src/main/java/com/rushcrew/order/domain/order/OrderRepository.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.order;
+
+public interface OrderRepository {
+}
