+++ order-service/src/main/java/com/rushcrew/order/domain/orderitem/OrderItem.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.orderitem;
+
+public class OrderItem {
+}
