+++ order-service/src/main/java/com/rushcrew/order/domain/orderitem/OrderItemRepository.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.orderitem;
+
+public interface OrderItemRepository {
+}
