+++ order-service/src/main/java/com/rushcrew/order/domain/reservation/OrderReservation.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.reservation;
+
+public class OrderReservation {
+}
