+++ order-service/src/main/java/com/rushcrew/order/domain/reservation/OrderReservationRepository.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.reservation;
+
+public interface OrderReservationRepository {
+}
