+++ order-service/src/main/java/com/rushcrew/order/domain/vo/OrderAmount.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.domain.vo;
+
+public class OrderAmount {
+}
