+++ order-service/src/main/java/com/rushcrew/order/infrastructure/client/feign/TimeDealStockClient.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.infrastructure.client.feign;
+
+public class TimeDealStockClient {
+}
