+++ order-service/src/main/java/com/rushcrew/order/infrastructure/config/RedisConfig.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.infrastructure.config;
+
+public class RedisConfig {
+}
