+++ order-service/src/main/java/com/rushcrew/order/infrastructure/messaging/kafka/producer/KafkaOrderProducer.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.infrastructure.messaging.kafka.producer;
+
+public class KafkaOrderProducer {
+}
