+++ order-service/src/main/java/com/rushcrew/order/infrastructure/persistence/OrderRepositoryImpl.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.infrastructure.persistence;
+
+public class OrderRepositoryImpl {
+}
