+++ order-service/src/main/java/com/rushcrew/order/presentation/controller/OrderController.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.presentation.controller;
+
+public class OrderController {
+}
