+++ order-service/src/main/java/com/rushcrew/order/presentation/dto/request/CreateOrderRequestDto.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.presentation.dto.request;
+
+public class CreateOrderRequestDto {
+}
