+++ order-service/src/main/java/com/rushcrew/order/presentation/dto/response/OrderResponseDto.java
@@ -1,0 +1,4 @@
+package com.rushcrew.order.presentation.dto.response;
+
+public class OrderResponseDto {
+}
