+++ order-service/src/main/java/com/rushcrew/order/presentation/exception/GlobalExceptionHandler.java
@@ -1,0 +1,7 @@
+package com.rushcrew.order.presentation.exception;
+
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+}
