--- order-service/src/main/java/com/rushdeal/order/OrderApplication.java
@@ -1,13 +1,0 @@
-package com.rushdeal.order;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderApplication.class, args);
-	}
-
-}
