--- order-service/src/main/java/com/rushdeal/order/application/event/producer/OrderEventPublisher.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.application.event.producer;
-
-public class OrderEventPublisher {
-}
