--- order-service/src/main/java/com/rushdeal/order/application/service/OrderService.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.application.service;
-
-public class OrderService {
-}
