--- order-service/src/main/java/com/rushdeal/order/domain/history/OrderHistory.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.domain.history;
-
-public class OrderHistory {
-}
