--- order-service/src/main/java/com/rushdeal/order/domain/history/OrderHistoryRepository.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.domain.history;
-
-public interface OrderHistoryRepository {
-}
