--- order-service/src/main/java/com/rushdeal/order/domain/order/Order.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.domain.order;
-
-public class Order {
-}
