--- order-service/src/main/java/com/rushdeal/order/domain/orderitem/OrderItem.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.domain.orderitem;
-
-public class OrderItem {
-}
