--- order-service/src/main/java/com/rushdeal/order/domain/orderitem/OrderItemRepository.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.domain.orderitem;
-
-public interface OrderItemRepository {
-}
