--- order-service/src/main/java/com/rushdeal/order/domain/reservation/OrderReservation.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.domain.reservation;
-
-public class OrderReservation {
-}
