--- order-service/src/main/java/com/rushdeal/order/domain/vo/OrderAmount.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.domain.vo;
-
-public class OrderAmount {
-}
