--- order-service/src/main/java/com/rushdeal/order/infrastructure/client/feign/TimeDealStockClient.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.infrastructure.client.feign;
-
-public class TimeDealStockClient {
-}
