--- order-service/src/main/java/com/rushdeal/order/infrastructure/config/KafkaConfig.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.infrastructure.config;
-
-public class KafkaConfig {
-}
