--- order-service/src/main/java/com/rushdeal/order/infrastructure/config/RedisConfig.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.infrastructure.config;
-
-public class RedisConfig {
-}
