--- order-service/src/main/java/com/rushdeal/order/infrastructure/messaging/kafka/producer/KafkaOrderProducer.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.infrastructure.messaging.kafka.producer;
-
-public class KafkaOrderProducer {
-}
