--- order-service/src/main/java/com/rushdeal/order/infrastructure/persistence/OrderRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.infrastructure.persistence;
-
-public class OrderRepositoryImpl {
-}
