--- order-service/src/main/java/com/rushdeal/order/presentation/controller/OrderController.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.presentation.controller;
-
-public class OrderController {
-}
