--- order-service/src/main/java/com/rushdeal/order/presentation/dto/request/CreateOrderRequestDto.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.presentation.dto.request;
-
-public class CreateOrderRequestDto {
-}
