--- order-service/src/main/java/com/rushdeal/order/presentation/dto/response/OrderResponseDto.java
@@ -1,4 +1,0 @@
-package com.rushdeal.order.presentation.dto.response;
-
-public class OrderResponseDto {
-}
