--- order-service/src/main/java/com/rushdeal/order/presentation/exception/GlobalExceptionHandler.java
@@ -1,7 +1,0 @@
-package com.rushdeal.order.presentation.exception;
-
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-}
