+++ order-service/src/test/java/com/rushcrew/order/OrderApplicationTests.java
@@ -1,0 +1,13 @@
+package com.rushcrew.order;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OrderApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
