--- order-service/src/test/java/com/rushdeal/order/OrderApplicationTests.java
@@ -1,13 +1,0 @@
-package com.rushdeal.order;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OrderApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
